+++ src/Response/Model/LinkedFbInfo.php
@@ -7,14 +7,19 @@ use InstagramAPI\AutoPropertyMapper;
 /**
  * LinkedFbInfo.
  *
+ * @method FacebookUser getLinkedFbPage()
  * @method LinkedFbUser getLinkedFbUser()
+ * @method bool isLinkedFbPage()
  * @method bool isLinkedFbUser()
+ * @method $this setLinkedFbPage(FacebookUser $value)
  * @method $this setLinkedFbUser(LinkedFbUser $value)
+ * @method $this unsetLinkedFbPage()
  * @method $this unsetLinkedFbUser()
  */
 class LinkedFbInfo extends AutoPropertyMapper
 {
     const JSON_PROPERTY_MAP = [
         'linked_fb_user'          => 'LinkedFbUser',
+        'linked_fb_page'          => 'FacebookUser',
     ];
 }